+++ .gitignore
@@ -7,3 +7,4 @@
 /Web_Service/build/
 /StackExchangeclient/dist/
 /Web_Service/dist/
+/Identity_Service/build/
