+++ Identity_Service/nbproject/project.properties
@@ -35,6 +35,7 @@ dist.war=${dist.dir}/${war.name}
 endorsed.classpath=
 excludes=
 file.reference.json-simple-1.1.1.jar=../asset/json-simple-1.1.1.jar
+file.reference.json-simple-1.1.1.jar-1=D:\\Plugin\\json-simple-1.1.1.jar
 includes=**
 j2ee.compile.on.save=true
 j2ee.copy.static.files.on.save=true
@@ -49,7 +50,8 @@ j2ee.platform.wsit.classpath=
 j2ee.server.type=gfv3ee6
 jar.compress=false
 javac.classpath=\
-    ${file.reference.json-simple-1.1.1.jar}
+    ${file.reference.json-simple-1.1.1.jar}:\
+    ${file.reference.json-simple-1.1.1.jar-1}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.debug=true
