+++ StackExchangeclient/web/controller/logout.jsp
@@ -1,0 +1,17 @@
+<%-- 
+    Document   : logout
+    Created on : Nov 18, 2015, 8:59:20 PM
+    Author     : Luqman A. Siswanto
+--%>
+
+<%@page contentType="text/html" pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>JSP Page</title>
+  </head>
+  <body>
+    <h1>Hello World!</h1>
+  </body>
+</html>
