+++ StackExchangeclient/web/view/register.jsp
@@ -26,8 +26,9 @@
       </form>
       <p class="text-right"><a href="login.jsp" class="link">Login</a> | <a href="register.jsp" class="link">Register</a></p>
 
-      <br> <br>
+      <p>Fill in the form to register into Overflow 48.</p><br>
         <div class="small_container">
+          <h5>Register in Overflow 48</h5><br>
           <form id="login" action="../controller/register.jsp" method="POST">
           <input id="name" placeholder="Name" class="form" type="text" name="name" value="">
           <input id="email" placeholder="Email" class="form" type="text" name="email" value="">
